convert Pagination.js to typescript
